fix typos in README.md
Wouldn't have fixed this if it weren't for the misspelling of 'deprecated' into the unrelated 'depreciated', but also noticed that 'IMPORTANT' was spelled wrong, too.

Look around online, but in short:

> Depreciated : diminished in value over a period of time

and

> Deprecated : no longer recommended and may be removed in the future, or close to becoming obsolete (programming)

'Deprecated' is not really used in any other contexts anymore, and neither is 'depreciated'. Although in dictionaries they do share a usage, it's not the one you're going for.